Move control panel into a httprequestbegin pipeline, eliminating the need to change the web.config to install Synthesis
