+++ Packages/repositories.config
@@ -4,4 +4,5 @@
   <repository path="..\Source\Synthesis.Blade\packages.config" />
   <repository path="..\Source\Synthesis.Mvc\packages.config" />
   <repository path="..\Source\Synthesis.Tests\packages.config" />
+  <repository path="..\Source\Synthesis\packages.config" />
 </repositories>
