+++ Source/Synthesis/Synthesis.config
@@ -360,6 +360,17 @@
 				<indexConfiguration name="sitecore_master_index" />
 			</providers>
 		</synthesis>
+		<pipelines>
+			<httpRequestBegin>
+				<!-- 
+					This pipeline handler shows the Synthesis control panel. You can customize the URL the control panel lives at by changing the activationUrl. 
+					The activationUrl must be a URL that the Sitecore pipeline won't ignore (e.g. .aspx, .ashx, etc)
+				-->
+				<processor patch:after="*[@type='Sitecore.Pipelines.HttpRequest.UserResolver, Sitecore.Kernel']" type="Synthesis.ControlPanel.ControlPanelPipelineProcessor">
+					<param desc="activationUrl">/synthesis.aspx</param>
+				</processor>
+			</httpRequestBegin>
+		</pipelines>
 		<contentSearch>
 			<indexConfigurations>
 				<defaultLuceneIndexConfiguration>
